+++ package.json
@@ -1,11 +1,12 @@
 {
   "name": "todolist",
+  "homepage": "./",
   "version": "1.0.0",
   "description": "to do list",
   "main": "./src/index.tsx",
   "scripts": {
-    "start": "webpack-dev-server .",
-    "build": "Webpack .",
+    "start": "webpack server --config webpack.dev.config.js",
+    "build": "webpack --config webpack.prod.config.js",
     "lint": "eslint src/**/*/*.{ts,tsx} --quiet --fix"
   },
   "author": "",
@@ -33,7 +34,7 @@
     "typescript": "5.2.2",
     "typescript-plugin-css-modules": "5.0.1",
     "webpack": "5.88.2",
-    "webpack-bundle-analyzer": "^4.9.1",
+    "webpack-bundle-analyzer": "4.9.1",
     "webpack-cli": "5.1.4",
     "webpack-dev-server": "4.15.1"
   },
