+++ src/components/TodoEditModal/TodoEditModal.module.css
@@ -43,16 +43,14 @@
   text-decoration: none;
   cursor: pointer;
 }
-.dateField_label {
-  max-width: 250px;
-}
+
 label {
   font-weight: bold;
 }
 .textarea {
   width: 100%;
   padding: 10px 15px;
-  margin-top: 5px;
+  margin: 5px 0;
   box-sizing: border-box;
   border: 2px solid #ccc;
   border-radius: 4px;
@@ -60,17 +58,24 @@ label {
   font-size: 16px;
   resize: none;
 }
+
+.textarea:invalid,
+.datePicker:invalid {
+  border-color: red;
+  border-width: 2px;
+}
+
 .textarea:focus,
 .datePicker:focus {
   color: #212529;
   background-color: #fff;
-  border-color: #bdbdbd;
+  /* border-color: #bdbdbd; */
   outline: 0;
   box-shadow: 0 0 0 0.2rem rgba(158, 158, 158, 0.25);
 }
 .datePicker {
   padding: 10px 15px;
-  margin-top: 5px;
+  margin: 5px 0;
   box-sizing: border-box;
   border: 2px solid #ccc;
   border-radius: 4px;
@@ -78,7 +83,7 @@ label {
   font-size: 16px;
 }
 .error {
-  border-color: red;
+  color: red;
 }
 
 .saveButton {
@@ -107,4 +112,14 @@ label {
 }
 .cancelButton:hover {
   background-color: #c82333;
+}
+
+.datePicker:invalid + span::after {
+  content: '✖';
+  padding-left: 5px;
+}
+
+.datePicker:valid + span::after {
+  content: '✓';
+  padding-left: 5px;
 }
