+++ src/components/TodoEditModal/TodoEditModal.tsx
@@ -2,6 +2,7 @@ import React, { useEffect, useRef, useState } from 'react';
 
 import styles from './TodoEditModal.module.css';
 import { ITodoData } from '@/type/ITodoData';
+import { validateForm } from '@/utils/todoValidation';
 
 interface IProps {
   closeModal: () => void;
@@ -15,8 +16,10 @@ export const TodoEditModal = ({
   dataModal,
 }: IProps) => {
   const [todo, setTodo] = useState<ITodoData | undefined>();
-  const [errors, setErrors] = useState({ inputValue: false, inputDate: false });
-
+  const [errMsg, setErrMsg] = useState<{ [key: string]: string }>({
+    value: '',
+    date: '',
+  });
   const textareaRef = useRef<HTMLTextAreaElement | null>(null);
   const inputDateRef = useRef<HTMLInputElement | null>(null);
 
@@ -24,24 +27,32 @@ export const TodoEditModal = ({
     setTodo(dataModal);
   }, []);
 
-  const handleTodoSubmit = () => {
-    if (!inputDateRef.current.value) {
-      setErrors((prev) => ({ ...prev, inputDate: true }));
-      return;
-    }
-
-    if (!textareaRef.current.value) {
-      setErrors((prev) => ({ ...prev, inputValue: true }));
-      return;
-    }
-
-    submitModal({
-      ...todo,
-      date: new Date(inputDateRef.current.value),
-      value: textareaRef.current.value,
+  const handleChange = (
+    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+  ) => {
+    const obj = validateForm({
+      [e.target.name]: e.target.value,
     });
 
-    closeModal();
+    const keys = Object.keys(obj)[0];
+    const val = obj[keys];
+    if (val) {
+      setErrMsg({ ...errMsg, [keys]: val });
+    } else {
+      errMsg[keys] && setErrMsg({ ...errMsg, [keys]: '' });
+    }
+  };
+
+  const handleSubmit = () => {
+    if (Object.values(errMsg).every((value) => value === '')) {
+      submitModal({
+        ...todo,
+        date: new Date(inputDateRef.current.value),
+        value: textareaRef.current.value,
+      });
+
+      closeModal();
+    }
   };
 
   return (
@@ -59,42 +70,40 @@ export const TodoEditModal = ({
             Value:
             <br />
             <textarea
-              className={
-                errors.inputValue
-                  ? `${styles.textarea} ${styles.error}`
-                  : styles.textarea
-              }
+              name="value"
+              rows={4}
+              className={styles.textarea}
               defaultValue={todo?.value}
+              onChange={(e) => handleChange(e)}
+              required={true}
+              minLength={3}
               ref={textareaRef}
-              onFocus={() => {
-                errors.inputValue &&
-                  setErrors((prev) => ({ ...prev, inputValue: false }));
-              }}
             />
+            {errMsg.value && <div className={styles.error}>{errMsg.value}</div>}
           </label>
 
-          <label className={styles.dateField_label}>
+          <label>
             Date:
             <br />
-            <input
-              className={
-                errors.inputDate
-                  ? `${styles.datePicker} ${styles.error}`
-                  : styles.datePicker
-              }
-              type="datetime-local"
-              step="1"
-              defaultValue={dataModal.date.toISOString().split('.')[0]}
-              ref={inputDateRef}
-              onFocus={() => {
-                errors.inputDate &&
-                  setErrors((prev) => ({ ...prev, inputDate: false }));
-              }}
-            />
+            <div>
+              <input
+                className={styles.datePicker}
+                type="datetime-local"
+                step="1"
+                name="date"
+                min="1986-01-01T00:00"
+                onChange={(e) => handleChange(e)}
+                defaultValue={dataModal.date.toISOString().split('.')[0]}
+                required={true}
+                ref={inputDateRef}
+              />
+              <span className="validity"></span>
+            </div>
+            {errMsg.date && <div className={styles.error}>{errMsg.date}</div>}
           </label>
         </div>
 
-        <button onClick={handleTodoSubmit} className={styles.saveButton}>
+        <button onClick={handleSubmit} className={styles.saveButton}>
           Save
         </button>
         <button onClick={closeModal} className={styles.cancelButton}>
