--- src/components/TodoEditModal1/TodoEditModal1.module.css
@@ -1,125 +1,0 @@
-.modal {
-  display: block;
-  position: fixed;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.7);
-}
-.modalContent {
-  background-color: #fefefe;
-  margin: 30px auto;
-  padding: 20px;
-  border: 1px solid #888;
-  max-width: 400px;
-  border-radius: 10px;
-}
-.header {
-  display: flex;
-  justify-content: center;
-}
-.title {
-  font-size: 24px;
-  font-weight: bold;
-}
-.editSection {
-  display: flex;
-  padding: 20px;
-  justify-content: space-around;
-  flex-direction: column;
-  gap: 20px;
-}
-.close {
-  color: #aaa;
-  float: right;
-  font-size: 28px;
-  font-weight: bold;
-}
-.close:hover,
-.close:focus {
-  color: black;
-  text-decoration: none;
-  cursor: pointer;
-}
-
-label {
-  font-weight: bold;
-}
-.textarea {
-  width: 100%;
-  padding: 10px 15px;
-  margin: 5px 0;
-  box-sizing: border-box;
-  border: 2px solid #ccc;
-  border-radius: 4px;
-  background-color: #f8f8f8;
-  font-size: 16px;
-  resize: none;
-}
-
-.textarea:invalid,
-.datePicker:invalid {
-  border-color: red;
-  border-width: 2px;
-}
-
-.textarea:focus,
-.datePicker:focus {
-  color: #212529;
-  background-color: #fff;
-  /* border-color: #bdbdbd; */
-  outline: 0;
-  box-shadow: 0 0 0 0.2rem rgba(158, 158, 158, 0.25);
-}
-.datePicker {
-  padding: 10px 15px;
-  margin: 5px 0;
-  box-sizing: border-box;
-  border: 2px solid #ccc;
-  border-radius: 4px;
-  background-color: #f8f8f8;
-  font-size: 16px;
-}
-.error {
-  color: red;
-}
-
-.saveButton {
-  background-color: #28a745;
-  color: #fff;
-  border: none;
-  border-radius: 4px;
-  padding: 10px 15px;
-  margin-right: 10px;
-  cursor: pointer;
-  font-size: 16px;
-  transition: background-color 0.3s ease;
-}
-.saveButton:hover {
-  background-color: #218838;
-}
-.cancelButton {
-  background-color: #dc3545;
-  color: #fff;
-  border: none;
-  border-radius: 4px;
-  padding: 10px 15px;
-  cursor: pointer;
-  font-size: 16px;
-  transition: background-color 0.3s ease;
-}
-.cancelButton:hover {
-  background-color: #c82333;
-}
-
-.datePicker:invalid + span::after {
-  content: '✖';
-  padding-left: 5px;
-}
-
-.datePicker:valid + span::after {
-  content: '✓';
-  padding-left: 5px;
-}
