--- src/components/index.ts
@@ -1,5 +1,0 @@
-export * from './TodoCreate/TodoCreate';
-export * from './TodoLists/TodoLists';
-export * from './Loading/Loading';
-export * from './Form/MyForm';
-export * from './List/List';
