+++ src/utils/todoValidation.ts
@@ -1,28 +1,37 @@
-interface FormData {
-  value: string;
-  date: string;
+interface IProps {
+  [key: string]: string;
 }
 
-export const validateForm = (formData: FormData) => {
-  const errors: Partial<FormData> = {};
+export const validateForm = (element: IProps) => {
+  const errors: {
+    [key: string]: string;
+  } = {};
 
-  for (const field in formData) {
+  for (const field in element) {
+    errors[field] = '';
+
     switch (field) {
       case 'value':
-        if (!formData[field].trim()) {
-          errors[field] = 'Text field is required.';
+        if (!element[field].trim()) {
+          errors[field] = 'Value field is required.';
           break;
         }
 
-        if (formData[field].length < 3) {
+        if (element[field].length < 3) {
           errors[field] = 'Value should be at least 3 characters.';
           break;
         }
         break;
 
       case 'date':
-        if (!formData[field]) {
+        if (!element[field]) {
           errors[field] = 'Date field is required';
+          break;
+        }
+
+        if (new Date(element[field]) < new Date('1986-01-01T00:00')) {
+          errors[field] = 'Date cannot be less than 1986';
+          break;
         }
         break;
     }
